Auto-copy config rather than mount
Why:

We cannot mount the .config inside the users home folder since the home
folder is already a mounted point. To bind folders with Singularity the
latter one needs to already exist in the container, hence we cannot bind
with folders inside another binded folder. To overcome this, we can copy
the current .config folder inside the binded home folder. Since we know
the home folder that will be bind, we know where to copy the .config.